requirements.txt
python-dotenv
discord.py
openai
wolframalpha
aiohttp